+++ src/main/java/app/slicequeue/sq_board/article/command/application/service/UpdateArticleService.java
@@ -1,10 +1,9 @@
-package app.slicequeue.sq_board.article.command.application;
+package app.slicequeue.sq_board.article.command.application.service;
 
 import app.slicequeue.common.exception.NotFoundException;
 import app.slicequeue.sq_board.article.command.domain.Article;
 import app.slicequeue.sq_board.article.command.domain.ArticleId;
 import app.slicequeue.sq_board.article.command.domain.ArticleRepository;
-import app.slicequeue.sq_board.article.command.domain.dto.DeleteArticleCommand;
 import app.slicequeue.sq_board.article.command.domain.dto.UpdateArticleCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,14 +12,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class DeleteArticleService {
+public class UpdateArticleService {
 
     private final ArticleRepository articleRepository;
 
-    public Article deleteArticle(DeleteArticleCommand command) {
+    public Article updateArticle(UpdateArticleCommand command) {
         Article article = articleRepository.findByArticleIdAndDeletedAtIsNull(command.articleId())
                 .orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다."));
-        article.delete();
+        article.update(command);
         return articleRepository.save(article);
     }
 }
