+++ src/main/java/app/slicequeue/sq_board/article/command/application/usecase/DeleteArticleUseCase.java
@@ -1,10 +1,11 @@
-package app.slicequeue.sq_board.article.command.application;
+package app.slicequeue.sq_board.article.command.application.usecase;
 
 import app.slicequeue.common.exception.NotFoundException;
+import app.slicequeue.sq_board.article.command.application.service.DeleteArticleService;
+import app.slicequeue.sq_board.article.command.application.service.SummarizeBoardArticleCountService;
 import app.slicequeue.sq_board.article.command.domain.Article;
 import app.slicequeue.sq_board.article.command.domain.ArticleId;
-import app.slicequeue.sq_board.article.command.domain.dto.BoardArticleCountCommand;
-import app.slicequeue.sq_board.article.command.domain.dto.CreateArticleCommand;
+import app.slicequeue.sq_board.article.command.domain.dto.DeleteArticleCommand;
 import app.slicequeue.sq_board.common.util.ConstraintViolationUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -15,12 +16,12 @@ import java.util.concurrent.atomic.AtomicReference;
 
 @Service
 @RequiredArgsConstructor
-public class CreateArticleUseCase {
+public class DeleteArticleUseCase {
 
-    private final CreateArticleService createArticleService;
+    private final DeleteArticleService deleteArticleService;
     private final SummarizeBoardArticleCountService summarizeBoardArticleCountService;
 
-    public ArticleId execute(CreateArticleCommand command) {
+    public ArticleId execute(DeleteArticleCommand command) {
         try {
             return executeWithTransaction(command);
         } catch (DataIntegrityViolationException e) {
@@ -32,13 +33,14 @@ public class CreateArticleUseCase {
     }
 
     @Transactional
-    public ArticleId executeWithTransaction(CreateArticleCommand command) {
+    public ArticleId executeWithTransaction(DeleteArticleCommand command) {
         AtomicReference<Article> article = new AtomicReference<>();
-        summarizeBoardArticleCountService.increaseWithLock(() -> {
-            Article created = createArticleService.createArticle(command);
-            article.set(created);
-            return created;
-        });
+        summarizeBoardArticleCountService.decreaseWithLock(
+                () -> {
+                    Article deleted = deleteArticleService.deleteArticle(command);
+                    article.set(deleted);
+                    return deleted;
+                });
         return article.get().getArticleId();
     }
 }
